update item bigdecimal attribute

+++ src/packages/dao/VendingMachineItem.java
@@ -1,12 +1,14 @@
 package packages.dao;
 
+import java.math.BigDecimal;
+
 public class VendingMachineItem {
     // attribute name will be key in the vending machine
     private String name;
-    private int cost;
+    private BigDecimal cost;
     private int qty;
 
-    public VendingMachineItem(String name, int cost, int qty) {
+    public VendingMachineItem(String name, BigDecimal cost, int qty) {
         this.name = name;
         this.cost = cost;
         this.qty = qty;
@@ -16,7 +18,7 @@ public class VendingMachineItem {
         return name;
     }
 
-    public int getCost() {
+    public BigDecimal getCost() {
         return cost;
     }
